requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
orjson==3.9.12
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
PyJWT==2.8.0
pytest==8.0.0
python-dotenv==1.0.1
python-multipart==0.0.7
PyYAML==6.0.1
setuptools==68.2.2
sniffio==1.3.0
SQLAlchemy==2.0.27
sqlmodel==0.0.16
starlette==0.35.1
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.41.2
